decomp(laneB): nomatch near-pair vein +35 strict (17,366→17,401)
Lane-B campaign on the identification-nomatch residue: new near-pair
scanner (union-reloc-masked word hamming, size/reloc-count gated) ranked
4,420 candidates; ~60 real targets dispatched to 8 crack agents.

+35 = 10 free map-entry flips (reloc-normalization-absorbed) + 25 source
cracks: Object.h ObjVector/ObjList op>>(BinStreamRev&) return bs (4 insts),
SfxInst mMoggClips ObjPtrList port, CameraShot.h mCrowdStateOverride order
(CamShot trio + 2 funclets), Campaign mispair→IsUserOnLastCampaignLevel
(+3 funclets), fabsf, named-local Symbol return slots, DC3-only init
strips, KeylessHash::Find over-carve merge (symbols.txt), + siblings.

Denylist: 0x82685750 (PracticeSection dtor ICF mispair).
Verified: independent clean-worktree whole-binary A/B at HEAD —
GAINED 35 / LOST 0.

+++ src/band3/game/Singer.cpp
@@ -423,7 +423,7 @@ void Singer::UpdatePitchDeviation(float pitch) {
     float mean = mPitchDeviationMean;
     float newMean = mean + (pitch - mean) / (float)count;
     mPitchDeviationMean = newMean;
-    mPitchDeviationDev = dev + (std::fabs(pitch - newMean) - dev) / (float)count;
+    mPitchDeviationDev = dev + (fabsf(pitch - newMean) - dev) / (float)count;
 }
 
 float Singer::GetPartPercentage(int part) const {
